refactor: use matvis insteaf of vis_cpu
