+++ .pre-commit-config.yaml
@@ -32,14 +32,16 @@ repos:
           - flake8-bugbear
           - flake8-comprehensions
           - flake8-print
--   repo: https://github.com/psf/black
-    rev: 23.11.0
-    hooks:
+
+- repo: https://github.com/psf/black-pre-commit-mirror
+  rev: 23.11.0
+  hooks:
     - id: black
--   repo: https://github.com/pre-commit/pygrep-hooks
-    rev: v1.10.0
-    hooks:
-      - id: rst-backticks
+
+- repo: https://github.com/pre-commit/pygrep-hooks
+  rev: v1.10.0
+  hooks:
+    - id: rst-backticks
 
 - repo: https://github.com/PyCQA/isort
   rev: 5.12.0
