+++ hera_sim/tests/testdata/hera-sim-vis-config/matvis_gpu.yaml
@@ -1,5 +1,6 @@
-simulator: VisCPU
+simulator: MatVis
 precision: 2
 ref_time: mean
 correct_source_positions: true
+use_gpu: true
 check_antenna_conjugation: false
